Creates MapGraph superclass and restructes the code
